requirements.txt
streamlit
google-generativeai
pandas
yfinance
matplotlib
plotly